TASK: Improve deferred persistent resource implementation
- use a static factory instead of constructor with optional argument
- prevent ResourceRepository from accepting deferred resources
- replace `getSource()` with `isDeferred()` getter, since you should never get the source without detaching

+++ Neos.Flow/Classes/ResourceManagement/ResourceRepository.php
@@ -74,6 +74,9 @@ class ResourceRepository extends Repository
      */
     public function add($object)
     {
+        if ($object instanceof PersistentResource && $object->isDeferred()) {
+            throw new IllegalObjectTypeException('You are trying to persist a deferred resource. Please use the ResourceManager instead.', 1587299343);
+        }
         $this->persistenceManager->allowObject($object);
         if ($this->removedResources->contains($object)) {
             $this->removedResources->detach($object);
